+++ NEWS.md
@@ -1,3 +1,9 @@
+# scplot (development version)
+
+- New function `add_line()`: Allows to add ab lines, horizontal or vertical lines.
+- "trendA theil-sen" now represents theil-sen regresssion line (not Siegel)
+- New theme "illustration"
+
 # scplot 0.3.3
 
 - moved `scan`dependency to `scan` import.